requirements.txt
transformers
gradio
python-dotenv
torch